Re-write PcsxrMemoryObject in Swift.

+++ macosx/Pcsxr-QL/GeneratePreviewForURL.m
@@ -6,6 +6,7 @@
 #import <Cocoa/Cocoa.h>
 //#include "nopic.h"
 #import "PSXMemEnumerator.h"
+#import "Pcsxr_QL-Swift.h"
 
 /* -----------------------------------------------------------------------------
    Generate a preview for file
@@ -118,7 +119,7 @@ static OSStatus GeneratePreviewForMemCard(void *thisInterface, QLPreviewRequestR
 			NSDictionary *imgProps = @{(NSString *)kQLPreviewPropertyAttachmentDataKey: pngData,
 									   (NSString *)kQLPreviewPropertyMIMETypeKey: @"image/png"};
 			NSString *imgName = [[@(i++) stringValue] stringByAppendingPathExtension:@"png"];
-			[htmlStr appendFormat:@"\t\t\t<tr><td><img src=\"cid:%@\"></td> <td>%@</td> <td>%i</td></tr>\n", imgName, obj.name, obj.blockSize];
+			[htmlStr appendFormat:@"\t\t\t<tr><td><img src=\"cid:%@\"></td> <td>%@</td> <td>%li</td></tr>\n", imgName, obj.name, (long)obj.blockSize];
 			htmlDict[imgName] = imgProps;
 			continue;
 		}
@@ -135,7 +136,7 @@ static OSStatus GeneratePreviewForMemCard(void *thisInterface, QLPreviewRequestR
 		NSDictionary *imgProps = @{(NSString *)kQLPreviewPropertyAttachmentDataKey: gifData,
 								   (NSString *)kQLPreviewPropertyMIMETypeKey: @"image/gif"};
 		NSString *imgName = [[@(i++) stringValue] stringByAppendingPathExtension:@"gif"];
-		[htmlStr appendFormat:@"\t\t\t<tr><td><img src=\"cid:%@\"></td> <td>%@</td> <td>%i</td></tr>\n", imgName, obj.name, obj.blockSize];
+		[htmlStr appendFormat:@"\t\t\t<tr><td><img src=\"cid:%@\"></td> <td>%@</td> <td>%li</td></tr>\n", imgName, obj.name, (long)obj.blockSize];
 		htmlDict[imgName] = imgProps;
 	}
